requirements.txt
openai>=1,<2
python-dotenv>=1.0.1
